layout/base/tests/test_bug990340.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=990340
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 990340</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 990340 **/

  function testbug990340() {
    ok(document.querySelector('#bug990340 span').clientHeight < 100,
       "'height' is in transition")
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=990340">Mozilla Bug 990340</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<style type="text/css">

#bug990340::before {
    content: ":before";
}

#bug990340 span {
    display: inline-block;
    border:1px solid blue;
    height: 20px;
    transition: height 30s;
}

#bug990340.s span {
    height: 100px;
}
</style>
<div id="bug990340" style="overflow:scroll">
    <span>Transition height</span>
</div>
</pre>

<script>
document.body.offsetHeight;
document.querySelector('#bug990340').className='s';
testbug990340()
</script>

</body>
</html>
